+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 *Minor*
 
 * Allowed `start` to work when the game has ended but `end` hasn't been called. ([#15](https://github.com/hjwylde/werewolf/issues/15))
+* Added `quit` command. ([#13](https://github.com/hjwylde/werewolf/issues/13))
 
 #### v0.2.0.0